Completes rehaul and generalization of expression indexer.
Expression indexer (formerly the fermata indexer) now returns the
kern representation of music21 expressions. It also aggregates
these results so if a note has more than one expression on it, it
will return a multiple-character string. Currently only fermatas
and mordents are supported, but more expressions can be added
easily.